+++ src/components/storysList/Card.css
